ai_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sentence-transformers>=2.2.2
numpy>=1.24.2
openai>=0.27.0
python-dotenv>=1.0.0